Refactor admin UI: slideouts & pagination
Introduce slideout panels and reusable widgets, simplify admin flows and tweak dark theme.

- Add new widgets: slideout_panel.dart, pagination.dart, permission_matrix.dart.
- Replace modal/dialog and route-based group/user/role detail flows with slideout-based UI; refactor AdminGroupsScreen to open a create slideout and group detail slideout, and convert create dialog into an inline slideout flow that shows the detail view after creation.
- Remove separate admin detail screens and their router routes (user/group/role detail screens and related route imports were removed from router.dart).
- Update AdminEventsScreen to include a section header and use LumaPagination for page navigation; simplify pagination layout and loading indicator handling.
- AdminInvitesScreen started switching to slideout usage.
- Adjust dark theme palette and surface colors for a more neutral gray appearance (several color constants and theme uses updated).

These changes consolidate admin interactions into slideout panels, add a permission matrix and pagination widget for reuse, and clean up routing by removing detail routes in favor of slideouts.
