+++ src/luma-web/lib/screens/admin/admin_groups_screen.dart
@@ -1,10 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
 
 import '../../models/group.dart';
+import '../../models/custom_role.dart';
+import '../../models/user.dart';
 import '../../services/user_service.dart';
 import '../../theme/tokens.dart';
 import '../../widgets/perm_button.dart';
+import '../../widgets/permission_matrix.dart';
+import '../../widgets/slideout_panel.dart';
 
 class AdminGroupsScreen extends StatefulWidget {
   final UserService userService;
@@ -35,18 +38,36 @@ class _AdminGroupsScreenState extends State<AdminGroupsScreen> {
     }
   }
 
-  void _showCreateDialog() {
-    showDialog(
+  void _showCreateGroupSlideout() {
+    final titleNotifier = ValueNotifier('Create group');
+    showSlideoutPanel(
       context: context,
-      builder: (_) => _CreateGroupDialog(
+      titleNotifier: titleNotifier,
+      bodyBuilder: (_) => _CreateGroupContent(
         userService: widget.userService,
-        onCreated: _load,
+        titleNotifier: titleNotifier,
+        onCreated: (group) {
+          _load();
+          return _GroupDetailContent(
+            group: group,
+            userService: widget.userService,
+            onChanged: _load,
+          );
+        },
       ),
-    );
+    ).whenComplete(() => titleNotifier.dispose());
   }
 
-  void _showManageDialog(GroupRecord group) {
-    context.go('/admin/groups/${group.id}', extra: group);
+  void _showGroupSlideout(GroupRecord group) {
+    showSlideoutPanel(
+      context: context,
+      title: group.name,
+      bodyBuilder: (_) => _GroupDetailContent(
+        group: group,
+        userService: widget.userService,
+        onChanged: _load,
+      ),
+    );
   }
 
   @override
@@ -60,18 +81,32 @@ class _AdminGroupsScreenState extends State<AdminGroupsScreen> {
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
             Row(
+              crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                Chip(
-                  label: Text(_groups == null ? 'Groups' : 'Groups (${_groups!.length})'),
-                  backgroundColor: colorScheme.secondaryContainer,
+                Expanded(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text('Groups',
+                          style: Theme.of(context)
+                              .textTheme
+                              .titleMedium
+                              ?.copyWith(fontWeight: FontWeight.w600)),
+                      const SizedBox(height: 2),
+                      Text(
+                        '${_groups?.length ?? 0} total \u00b7 Organize users into teams.',
+                        style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                              color: colorScheme.onSurfaceVariant),
+                      ),
+                    ],
+                  ),
                 ),
-                const Spacer(),
                 PermButton(
                   label: 'Create group',
                   filled: true,
                   enabled: widget.userService.canCreateGroup,
                   requiredPermission: 'group:create',
-                  onPressed: _showCreateDialog,
+                  onPressed: _showCreateGroupSlideout,
                 ),
               ],
             ),
@@ -105,7 +140,7 @@ class _AdminGroupsScreenState extends State<AdminGroupsScreen> {
                         label: 'Manage',
                         enabled: widget.userService.canEditGroup,
                         requiredPermission: 'group:rename',
-                        onPressed: () => _showManageDialog(g),
+                        onPressed: () => _showGroupSlideout(g),
                       ),
                     );
                   },
@@ -118,22 +153,444 @@ class _AdminGroupsScreenState extends State<AdminGroupsScreen> {
   }
 }
 
-// ── Create group dialog ──────────────────────────────────────────────────────
+// ── Group detail content (slideout body) ─────────────────────────────────────
 
-class _CreateGroupDialog extends StatefulWidget {
+class _GroupDetailContent extends StatefulWidget {
+  final GroupRecord group;
   final UserService userService;
-  final VoidCallback onCreated;
-  const _CreateGroupDialog({required this.userService, required this.onCreated});
+  final VoidCallback onChanged;
+
+  const _GroupDetailContent({
+    required this.group,
+    required this.userService,
+    required this.onChanged,
+  });
 
   @override
-  State<_CreateGroupDialog> createState() => _CreateGroupDialogState();
+  State<_GroupDetailContent> createState() => _GroupDetailContentState();
 }
 
-class _CreateGroupDialogState extends State<_CreateGroupDialog> {
+class _GroupDetailContentState extends State<_GroupDetailContent> {
+  late final TextEditingController _nameCtrl;
+  late final TextEditingController _descCtrl;
+  GroupRecord? _detail;
+  List<AdminUserRecord>? _allUsers;
+  List<GroupRecord>? _allGroups;
+  List<CustomRoleRecord>? _allRoles;
+  bool _loading = true;
+  String? _error;
+
+  @override
+  void initState() {
+    super.initState();
+    _nameCtrl = TextEditingController(text: widget.group.name);
+    _descCtrl = TextEditingController(text: widget.group.description ?? '');
+    _loadDetail();
+  }
+
+  @override
+  void dispose() {
+    _nameCtrl.dispose();
+    _descCtrl.dispose();
+    super.dispose();
+  }
+
+  Future<void> _loadDetail() async {
+    setState(() { _loading = true; _error = null; });
+    try {
+      final results = await Future.wait([
+        widget.userService.getGroup(widget.group.id),
+        widget.userService.listAdminUsers(),
+        widget.userService.listGroups(),
+        widget.userService.listCustomRoles(),
+      ]);
+      if (mounted) {
+        final detail = results[0] as GroupRecord;
+        setState(() {
+          _detail = detail;
+          _nameCtrl.text = detail.name;
+          _descCtrl.text = detail.description ?? '';
+          _allUsers = results[1] as List<AdminUserRecord>;
+          _allGroups = (results[2] as List<GroupRecord>)
+              .where((g) => g.id != widget.group.id)
+              .toList();
+          _allRoles = results[3] as List<CustomRoleRecord>;
+          _loading = false;
+        });
+      }
+    } catch (e) {
+      if (mounted) {
+        setState(() {
+          _error = 'Could not load group details. Please try again.';
+          _loading = false;
+        });
+      }
+    }
+  }
+
+  Future<void> _rename() async {
+    if (_nameCtrl.text.trim().isEmpty) return;
+    final desc = _descCtrl.text.trim();
+    try {
+      await widget.userService.renameGroup(
+        widget.group.id,
+        _nameCtrl.text.trim(),
+        description: desc.isEmpty ? null : desc,
+        clearDescription: desc.isEmpty,
+      );
+      await _loadDetail();
+      widget.onChanged();
+    } catch (e) {
+      if (mounted) _showError(e.toString());
+    }
+  }
+
+  Future<void> _delete() async {
+    final ok = await showDialog<bool>(
+      context: context,
+      builder: (_) => AlertDialog(
+        title: const Text('Delete group'),
+        content: const Text('Are you sure? This cannot be undone.'),
+        actions: [
+          TextButton(
+              onPressed: () => Navigator.of(context).pop(false),
+              child: const Text('Cancel')),
+          FilledButton(
+              onPressed: () => Navigator.of(context).pop(true),
+              child: const Text('Delete')),
+        ],
+      ),
+    );
+    if (ok != true || !mounted) return;
+    try {
+      await widget.userService.deleteGroup(widget.group.id);
+      widget.onChanged();
+      if (mounted) Navigator.of(context).pop();
+    } catch (e) {
+      if (mounted) _showError(e.toString());
+    }
+  }
+
+  void _showError(String msg) {
+    final display = msg.replaceFirst('Exception: ', '');
+    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+      content: Text(
+          display.isEmpty ? 'Something went wrong. Please try again.' : display),
+      backgroundColor: Theme.of(context).colorScheme.error,
+    ));
+  }
+
+  Future<void> _addUserMember(String userId) async {
+    try {
+      await widget.userService.addGroupMember(widget.group.id, 'user', userId);
+      await _loadDetail();
+      widget.onChanged();
+    } catch (e) {
+      if (mounted) _showError(e.toString());
+    }
+  }
+
+  Future<void> _addGroupMember(String groupId) async {
+    try {
+      await widget.userService.addGroupMember(widget.group.id, 'group', groupId);
+      await _loadDetail();
+      widget.onChanged();
+    } catch (e) {
+      if (mounted) _showError(e.toString());
+    }
+  }
+
+  Future<void> _removeMember(GroupMemberRecord m) async {
+    final label = _memberLabel(m);
+    final confirmed = await showDialog<bool>(
+      context: context,
+      builder: (ctx) => AlertDialog(
+        title: const Text('Remove member?'),
+        content: Text('Remove "$label" from this group?'),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.of(ctx).pop(false),
+            child: const Text('Cancel'),
+          ),
+          FilledButton(
+            onPressed: () => Navigator.of(ctx).pop(true),
+            child: const Text('Remove'),
+          ),
+        ],
+      ),
+    );
+    if (confirmed != true) return;
+    try {
+      await widget.userService.removeGroupMember(widget.group.id, m.memberType, m.memberId);
+      await _loadDetail();
+      widget.onChanged();
+    } catch (e) {
+      if (mounted) _showError('Could not remove member. Please try again.');
+    }
+  }
+
+  Future<void> _assignRole(String roleId) async {
+    try {
+      await widget.userService.assignRoleToGroup(widget.group.id, roleId);
+      await _loadDetail();
+      widget.onChanged();
+    } catch (e) {
+      if (mounted) _showError(e.toString());
+    }
+  }
+
+  Future<void> _removeRole(String roleId) async {
+    try {
+      await widget.userService.removeRoleFromGroup(widget.group.id, roleId);
+      await _loadDetail();
+      widget.onChanged();
+    } catch (e) {
+      if (mounted) _showError(e.toString());
+    }
+  }
+
+  String _memberLabel(GroupMemberRecord m) {
+    if (m.memberType == 'user') {
+      final u = _allUsers?.where((u) => u.id == m.memberId).firstOrNull;
+      return u?.displayName ?? m.memberId;
+    } else {
+      final allG = (_allGroups ?? []) + [widget.group];
+      final found = allG.where((g) => g.id == m.memberId).firstOrNull;
+      return found?.name ?? m.memberId;
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final colorScheme = Theme.of(context).colorScheme;
+
+    if (_loading) {
+      return const Center(child: CircularProgressIndicator());
+    }
+
+    if (_error != null) {
+      return Center(
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Text(_error!, style: TextStyle(color: colorScheme.error)),
+            const SizedBox(height: 12),
+            FilledButton(onPressed: _loadDetail, child: const Text('Retry')),
+          ],
+        ),
+      );
+    }
+
+    final detail = _detail!;
+    final isSystem = detail.isSystem;
+    final noMemberControl = detail.noMemberControl;
+    final canDelete = detail.memberCount == 0 && !isSystem;
+
+    return SingleChildScrollView(
+      padding: const EdgeInsets.all(24),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          if (isSystem)
+            Padding(
+              padding: const EdgeInsets.only(bottom: 12),
+              child: const _SystemBadge(),
+            ),
+
+          // ── Name + Description ───────────────────────
+          Text('Name', style: Theme.of(context).textTheme.labelLarge),
+          const SizedBox(height: 8),
+          Row(
+            children: [
+              Expanded(
+                child: TextField(
+                  controller: _nameCtrl,
+                  enabled: !isSystem,
+                  decoration: const InputDecoration(
+                    isDense: true,
+                    border: OutlineInputBorder(),
+                  ),
+                ),
+              ),
+              const SizedBox(width: 8),
+              OutlinedButton(
+                  onPressed: isSystem ? null : _rename,
+                  child: const Text('Save')),
+            ],
+          ),
+          const SizedBox(height: 8),
+          TextField(
+            controller: _descCtrl,
+            enabled: !isSystem,
+            decoration: const InputDecoration(
+              labelText: 'Description (optional)',
+              isDense: true,
+              border: OutlineInputBorder(),
+            ),
+            maxLines: 2,
+          ),
+          const SizedBox(height: 6),
+          Text(
+            widget.group.id,
+            style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                  color: colorScheme.onSurfaceVariant.withAlpha(100),
+                  fontFamily: 'monospace',
+                ),
+          ),
+          const SizedBox(height: 12),
+          OutlinedButton.icon(
+            onPressed: canDelete ? _delete : null,
+            icon: const Icon(Icons.delete_outline, size: 16),
+            label: Text(isSystem
+                ? 'Cannot delete (system group)'
+                : canDelete
+                    ? 'Delete group'
+                    : 'Cannot delete (has members)'),
+            style: OutlinedButton.styleFrom(
+              foregroundColor: canDelete ? colorScheme.error : null,
+            ),
+          ),
+
+          const SizedBox(height: 8),
+          SwitchListTile(
+            contentPadding: EdgeInsets.zero,
+            title: const Text('Hide from directory search'),
+            subtitle: const Text(
+              'When enabled, this group will not appear in search results for non-admin users.',
+            ),
+            value: detail.hideFromSearch,
+            onChanged: isSystem
+                ? null
+                : (hide) async {
+                    try {
+                      await widget.userService
+                          .setGroupHideFromSearch(widget.group.id, hide: hide);
+                      await _loadDetail();
+                      widget.onChanged();
+                    } catch (e) {
+                      if (mounted) _showError(e.toString());
+                    }
+                  },
+          ),
+
+          const Divider(height: 40),
+
+          // ── Members ──────────────────────────────────
+          Text('Members', style: Theme.of(context).textTheme.labelLarge),
+          const SizedBox(height: 8),
+          if (noMemberControl)
+            Text(
+              'Membership is managed automatically by the system.',
+              style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                  color: colorScheme.onSurfaceVariant),
+            )
+          else if (detail.members.isEmpty)
+            const Text('No members')
+          else
+            ...detail.members.map((m) => ListTile(
+                  dense: true,
+                  leading: Icon(
+                    m.memberType == 'user'
+                        ? Icons.person_outline
+                        : Icons.group_outlined,
+                    size: 20,
+                  ),
+                  title: Text(_memberLabel(m)),
+                  subtitle: Text(m.memberType),
+                  trailing: IconButton(
+                    icon: const Icon(Icons.remove_circle_outline, size: 18),
+                    onPressed: () => _removeMember(m),
+                  ),
+                )),
+          if (!noMemberControl) ...[
+            const SizedBox(height: 8),
+            _AddMemberRow(
+              allUsers: _allUsers ?? [],
+              allGroups: _allGroups ?? [],
+              existingMembers: detail.members,
+              onAddUser: _addUserMember,
+              onAddGroup: _addGroupMember,
+            ),
+          ],
+
+          const Divider(height: 40),
+
+          // ── Roles ─────────────────────────────────────
+          Text('Assigned roles', style: Theme.of(context).textTheme.labelLarge),
+          const SizedBox(height: 8),
+          if (detail.roleIds.isEmpty)
+            const Text('No roles assigned')
+          else
+            ...detail.roleIds.map((rid) {
+              final role = _allRoles?.where((r) => r.id == rid).firstOrNull;
+              return ListTile(
+                dense: true,
+                leading: const Icon(Icons.shield_outlined, size: 20),
+                title: Text(role?.name ?? rid),
+                trailing: IconButton(
+                  icon: const Icon(Icons.remove_circle_outline, size: 18),
+                  onPressed: () => _removeRole(rid),
+                ),
+              );
+            }),
+          const SizedBox(height: 8),
+          _AddRoleRow(
+            allRoles: _allRoles ?? [],
+            assignedIds: detail.roleIds,
+            onAssign: _assignRole,
+          ),
+
+          const Divider(height: 40),
+
+          // ── Effective permissions ───────────────────────
+          Text('Effective permissions', style: Theme.of(context).textTheme.labelLarge),
+          const SizedBox(height: 4),
+          if (detail.roleIds.isEmpty)
+            Text(
+              'No roles assigned — no permissions granted.',
+              style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                    color: colorScheme.onSurfaceVariant,
+                  ),
+            )
+          else
+            PermissionMatrix(
+              permMap: resolveEffectivePermissions(
+                (_allRoles ?? [])
+                    .where((r) => detail.roleIds.contains(r.id))
+                    .toList(),
+              ),
+              readOnly: true,
+            ),
+          const SizedBox(height: 24),
+        ],
+      ),
+    );
+  }
+}
+
+// ── Create group slideout content ─────────────────────────────────────────────
+
+class _CreateGroupContent extends StatefulWidget {
+  final UserService userService;
+  final ValueNotifier<String> titleNotifier;
+  final Widget Function(GroupRecord group) onCreated;
+
+  const _CreateGroupContent({
+    required this.userService,
+    required this.titleNotifier,
+    required this.onCreated,
+  });
+
+  @override
+  State<_CreateGroupContent> createState() => _CreateGroupContentState();
+}
+
+class _CreateGroupContentState extends State<_CreateGroupContent> {
   final _nameCtrl = TextEditingController();
   final _descCtrl = TextEditingController();
   bool _saving = false;
   String? _error;
+  Widget? _detailView;
 
   @override
   void dispose() { _nameCtrl.dispose(); _descCtrl.dispose(); super.dispose(); }
@@ -143,11 +600,13 @@ class _CreateGroupDialogState extends State<_CreateGroupDialog> {
     setState(() { _saving = true; _error = null; });
     try {
       final desc = _descCtrl.text.trim();
-      await widget.userService.createGroup(
+      final group = await widget.userService.createGroup(
         _nameCtrl.text.trim(),
         description: desc.isEmpty ? null : desc,
       );
-      if (mounted) { Navigator.of(context).pop(); widget.onCreated(); }
+      if (!mounted) return;
+      widget.titleNotifier.value = group.name;
+      setState(() => _detailView = widget.onCreated(group));
     } catch (e) {
       if (mounted) setState(() { _error = e.toString().replaceFirst('Exception: ', ''); _saving = false; });
     }
@@ -155,39 +614,187 @@ class _CreateGroupDialogState extends State<_CreateGroupDialog> {
 
   @override
   Widget build(BuildContext context) {
-    return AlertDialog(
-      title: const Text('Create group'),
-      content: SizedBox(
-        width: 360,
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            TextField(
-              controller: _nameCtrl,
-              decoration: const InputDecoration(labelText: 'Group name'),
-              autofocus: true,
-              onSubmitted: (_) => _submit(),
-            ),
-            const SizedBox(height: 12),
-            TextField(
-              controller: _descCtrl,
-              decoration: const InputDecoration(labelText: 'Description (optional)'),
-              maxLines: 2,
-            ),
-            if (_error != null) ...[
-              const SizedBox(height: 8),
-              Text(_error!, style: TextStyle(color: Theme.of(context).colorScheme.error)),
-            ],
+    if (_detailView != null) return _detailView!;
+
+    final cs = Theme.of(context).colorScheme;
+
+    return SingleChildScrollView(
+      padding: const EdgeInsets.all(20),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: [
+          TextField(
+            controller: _nameCtrl,
+            decoration: const InputDecoration(labelText: 'Group name'),
+            autofocus: true,
+            onSubmitted: (_) => _submit(),
+          ),
+          const SizedBox(height: 12),
+          TextField(
+            controller: _descCtrl,
+            decoration: const InputDecoration(labelText: 'Description (optional)'),
+            maxLines: 2,
+          ),
+          if (_error != null) ...[
+            const SizedBox(height: 8),
+            Text(_error!, style: TextStyle(color: cs.error)),
           ],
-        ),
+          const SizedBox(height: 16),
+          FilledButton(
+            onPressed: _saving ? null : _submit,
+            child: _saving
+                ? const SizedBox(
+                    width: 16, height: 16,
+                    child: CircularProgressIndicator(strokeWidth: 2))
+                : const Text('Create group'),
+          ),
+        ],
       ),
-      actions: [
-        TextButton(onPressed: () => Navigator.of(context).pop(), child: const Text('Cancel')),
-        FilledButton(onPressed: _saving ? null : _submit, child: const Text('Create')),
+    );
+  }
+}
+
+// ── Add member row ────────────────────────────────────────────────────────────
+
+class _AddMemberRow extends StatefulWidget {
+  final List<AdminUserRecord> allUsers;
+  final List<GroupRecord> allGroups;
+  final List<GroupMemberRecord> existingMembers;
+  final void Function(String) onAddUser;
+  final void Function(String) onAddGroup;
+
+  const _AddMemberRow({
+    required this.allUsers,
+    required this.allGroups,
+    required this.existingMembers,
+    required this.onAddUser,
+    required this.onAddGroup,
+  });
+
+  @override
+  State<_AddMemberRow> createState() => _AddMemberRowState();
+}
+
+class _AddMemberRowState extends State<_AddMemberRow> {
+  String _type = 'user';
+  String? _selectedId;
+
+  @override
+  Widget build(BuildContext context) {
+    final existingIds = widget.existingMembers
+        .where((m) => m.memberType == _type)
+        .map((m) => m.memberId)
+        .toSet();
+
+    final items = _type == 'user'
+        ? widget.allUsers
+            .where((u) => !existingIds.contains(u.id))
+            .map((u) => DropdownMenuItem(value: u.id, child: Text(u.displayName)))
+            .toList()
+        : widget.allGroups
+            .where((g) => !existingIds.contains(g.id))
+            .map((g) => DropdownMenuItem(value: g.id, child: Text(g.name)))
+            .toList();
+
+    return Row(
+      children: [
+        DropdownButton<String>(
+          value: _type,
+          items: const [
+            DropdownMenuItem(value: 'user', child: Text('User')),
+            DropdownMenuItem(value: 'group', child: Text('Group')),
+          ],
+          onChanged: (v) =>
+              setState(() {
+                _type = v!;
+                _selectedId = null;
+              }),
+        ),
+        const SizedBox(width: 8),
+        Expanded(
+          child: DropdownButton<String>(
+            value: _selectedId,
+            hint: Text('Select $_type'),
+            isExpanded: true,
+            items: items,
+            onChanged: (v) => setState(() => _selectedId = v),
+          ),
+        ),
+        const SizedBox(width: 8),
+        IconButton(
+          icon: const Icon(Icons.add_circle_outline),
+          onPressed: _selectedId == null
+              ? null
+              : () {
+                  if (_type == 'user') {
+                    widget.onAddUser(_selectedId!);
+                  } else {
+                    widget.onAddGroup(_selectedId!);
+                  }
+                  setState(() => _selectedId = null);
+                },
+        ),
       ],
     );
   }
 }
+
+// ── Add role row ──────────────────────────────────────────────────────────────
+
+class _AddRoleRow extends StatefulWidget {
+  final List<CustomRoleRecord> allRoles;
+  final List<String> assignedIds;
+  final void Function(String) onAssign;
+
+  const _AddRoleRow({
+    required this.allRoles,
+    required this.assignedIds,
+    required this.onAssign,
+  });
+
+  @override
+  State<_AddRoleRow> createState() => _AddRoleRowState();
+}
+
+class _AddRoleRowState extends State<_AddRoleRow> {
+  String? _selectedId;
+
+  @override
+  Widget build(BuildContext context) {
+    final available = widget.allRoles
+        .where((r) => !widget.assignedIds.contains(r.id))
+        .toList();
+
+    return Row(
+      children: [
+        Expanded(
+          child: DropdownButton<String>(
+            value: _selectedId,
+            hint: const Text('Select role'),
+            isExpanded: true,
+            items: available
+                .map((r) =>
+                    DropdownMenuItem(value: r.id, child: Text(r.name)))
+                .toList(),
+            onChanged: (v) => setState(() => _selectedId = v),
+          ),
+        ),
+        const SizedBox(width: 8),
+        IconButton(
+          icon: const Icon(Icons.add_circle_outline),
+          onPressed: _selectedId == null
+              ? null
+              : () {
+                  widget.onAssign(_selectedId!);
+                  setState(() => _selectedId = null);
+                },
+        ),
+      ],
+    );
+  }
+}
+
+// ── System badge ─────────────────────────────────────────────────────────────
 
 class _SystemBadge extends StatelessWidget {
   const _SystemBadge();
